backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.3

# Database
supabase>=2.3.4
postgrest>=0.13.2
httpx>=0.24.0

# LiveKit
livekit
livekit-agents[silero,turn-detector]>=1.3.0
livekit-plugins-noise-cancellation>=0.2.0
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-api

# AI
anthropic==0.18.1

# Observability (Arize Phoenix)
arize-phoenix>=3.13.0
opentelemetry-api
opentelemetry-sdk
openinference-instrumentation-anthropic==0.1.0

# Utilities

requests==2.31.0
python-multipart==0.0.6
